refactor: Transform the "page of form" in a "component of form"
